fix data models

+++ weather-app/app/src/main/java/com/example/weatherapp/business/datamodel/Main.kt
@@ -2,4 +2,4 @@ package com.example.weatherapp.business.datamodel
 
 import com.google.gson.annotations.Expose
 
-data class Main(@Expose val temp: Int)
+data class Main(@Expose val temp: Double)
