+++ weather-app/app/src/main/java/com/example/weatherapp/business/datamodel/Wind.kt
@@ -2,4 +2,4 @@ package com.example.weatherapp.business.datamodel
 
 import com.google.gson.annotations.Expose
 
-data class Wind(@Expose val speed: Int, @Expose val deg: Int)
+data class Wind(@Expose val speed: Double, @Expose val deg: Double)
